+++ bin/extract_data_branch_from_pr.py
@@ -3,7 +3,7 @@
 import argparse
 import os
 
-from ecobalyse.github import extract_branch_name, get_github
+from ecobalyse_lib.github import extract_branch_name, get_github
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
