+++ packages/python/ecobalyse-lib/src/ecobalyse_lib/0001-feat-patch-homepage-link-and-inject-and-inject-versi.patch
